Move tests to new controller test
